backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
gunicorn==20.1.0
whitenoise==6.5.0
dotenv==0.9.9
Pillow==10.1.0
coverage==7.6.11
reportlab>=3.6.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-decouple==3.8
markdown==3.7
markdown==3.7
stripe==12.5.0
typing-extensions==4.15.0
